+++ core/src/com/ict1009/pokemanz/item/Coin.java
@@ -6,7 +6,7 @@ import com.ict1009.pokemanz.entity.Player;
 public class Coin extends Item {
     final private int value;
 
-    public Coin(World world, String textureLocation, float initialX, float initialY, int value) {
+    public Coin(World world, String textureLocation, int initialX, int initialY, int value) {
         super(world, textureLocation, initialX, initialY);
         this.value = value;
     }
